requirements.txt
requests
psycopg2-binary
pydantic>=2.0,<3
python-dotenv
google-cloud-secret-manager
functions-framework